% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_averages.R
\name{add_EMA.data.frame}
\alias{add_EMA.data.frame}
\alias{add_SMA.data.frame}
\title{add_EMA.data.frame}
\usage{
\method{add_EMA}{data.frame}(object, variable = "Close", n = 10, smooth = 2, na.rm = TRUE)

\method{add_SMA}{data.frame}(object, variable = "Close", n = 10, na.rm = TRUE)
}
\arguments{
\item{variable}{the name of the variable of which we would compute the moving average.}

\item{n}{integer, the number of length of the moving average.}

\item{smooth}{numeric, a parameter to use for smoothing the moving average.}

\item{na.rm}{logical,}
}
\value{
a tibble with one more column named as \code{EMA_} + \code{variable}, \verb{_n}.

a tibble with one more column named as \code{SMA_} + \code{variable}, \verb{_n}.
}
\description{
method for the class of objects \code{data.frame} for computing the Exponential Moving Average and adding to the
dataset in a tidy way.

method for the class of objects \code{data.frame} for computing the Simple Moving Average and adding to the
dataset in a tidy way.
}
\examples{
df <- data.frame(x = runif(1000, 0, 10))
df <- add_EMA(df, variable = "x", n = 10)


df <- data.frame(x = runif(1000, 0, 10))

# the new column will be named as "SMA_x_10"
df <- add_EMA(df, variable = "x", n = 10)

}
