% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_averages.R
\name{add_EMA}
\alias{add_EMA}
\title{add_EMA}
\usage{
add_EMA(object, ...)
}
\arguments{
\item{...}{other parameters.}
}
\description{
generic method for compute the Exponential Moving Average of an object.
}
