% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{Api}
\alias{Api}
\title{Api}
\usage{
Api(api = "blockchain", paths = NULL, query = NULL,
  config = list(), quiet = TRUE, json = FALSE)
}
\arguments{
\item{api}{character, api to use for importing data (cryptocompare)}

\item{paths}{vector, paths for the contruction of the link (see api documentation )}

\item{query}{list, query for the contruction of the link (see api documentation )}

\item{config}{configuration param \code{\link[httr]{config}},
see also \code{\link[httr]{authenticate}}, from httr package.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}

\item{json}{logical. When TRUE the response will be converted from json, else is converted in the base form
with the function \code{\link[httr]{content}}}
}
\description{
the base function for importing data from a web Api, it is scalable, if you want to add more
apis you have just to add the new base link inside the funcion in api.supported variable.
}
