% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyptocompare.R
\name{cc_all_exchanges}
\alias{cc_all_exchanges}
\title{cc_all_exchanges}
\usage{
cc_all_exchanges()
}
\value{
character vector
}
\description{
retrieve the names of all the exchanges available in cryptocompare Api.
}
\examples{
cc_all_exchanges()
}
