% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_blockchain_balance.R
\name{cc_blockchain_balance}
\alias{cc_blockchain_balance}
\title{cc_blockchain_balance}
\usage{
cc_blockchain_balance(from = NULL, to = NULL, api_key = NULL)
}
\arguments{
\item{from}{character or Date, the start date for importation.}

\item{to}{character or Date, the end date for importation.}

\item{api_key}{character}
}
\value{
tibble
}
\description{
historical balance for Bitcoin blockchain only. The output will be a tibble containing the total Bitcoin held and the number of wallet by different amounts range.
In order to access to this endpoint a \verb{free api key is needed}.
}
\details{
The output of the function will be a tibble with 6 columns:
\itemize{
\item \code{Date}: the observation date.
\item \code{Symbol}: the symbol of interest.
\item \code{From}: the minimum amount of Bitcoin held by some wallets.
\item \code{To}: the maximum amount of Bitcoin held by some wallets.
\item \code{Amount}: the total number of Bitcoin held by wallets having a quantity of Bitcoin between From and To.
\item \code{Wallets}: the total number of Wallets having a quantity of Bitcoin between From and To.
}
}
\examples{

# Api key is required
\dontrun{
yourapikey <- "yourapikey"
cc_blockchain_balance(from = "2017-01-01",
                      to = "2023-01-01",
                      api_key = yourapikey)
}
}
