% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_blockchain_historical.R
\name{cc_blockchain_historical}
\alias{cc_blockchain_historical}
\title{cc_blockchain_historical}
\usage{
cc_blockchain_historical(
  symbol = NULL,
  start = NULL,
  end = NULL,
  api_key = NULL
)
}
\arguments{
\item{symbol}{character, the symbol of interest.}

\item{start}{character or Date, the start date for importation.}

\item{end}{character or Date, the end date for importation.}

\item{api_key}{character}
}
\value{
A \code{tibble}.
}
\description{
Blockchain historical data, such as the number of transactions, active wallets or mining rewards are difficult to retrieve in a readable format. Cryptocompare simplify
substantially the procedure of download, aggregate and summaries the data for each day. With this function it is possible to obtain the historical information for most of the blockchain.
The data available for the Bitcoin's blockchain starts from 2017. In order to access to this function a \verb{free api key is needed}.
}
\examples{
# Api key is required
\dontrun{
yourapikey <- "yourapikey"

cc_blockchain_historical(symbol = "BTC", start = "2017-01-01",
                         end = "2023-01-01", api_key = yourapikey)

cc_blockchain_historical(symbol = "ETH", start = "2017-01-01",
                         end = "2023-01-01", api_key = yourapikey)

cc_blockchain_historical(symbol = "DOGE", start = "2017-01-01",
                         end = "2023-01-01", api_key = yourapikey)
}
}
