% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_options.R
\name{cc_category_options}
\alias{cc_category_options}
\title{cc_category_options}
\usage{
cc_category_options()
}
\value{
a character vector with all categories
}
\description{
visualize all news categories availables.
}
