% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_clean_env}
\alias{cc_clean_env}
\title{cc_clean_env}
\usage{
cc_clean_env()
}
\description{
: delete all elements in the environment cc_env.
}
\examples{
\donttest{cc_clean_env() }
}
\seealso{
\code{\link{cc_set_env}}
}
