% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_eth_staking.R
\name{cc_eth_staking}
\alias{cc_eth_staking}
\title{cc_eth_staking}
\usage{
cc_eth_staking(start = NULL, end = NULL, api_key = NULL)
}
\arguments{
\item{start}{character or Date, the starting date for importation.}

\item{end}{character or Date, the end date for importation.}

\item{api_key}{character}
}
\value{
tibble
}
\description{
Ethereum staking interest rate, from when the Beacon chain was active. In order to access to this endpoint a \verb{free api key is needed}.
}
\examples{

# Api key is required

yourapikey <- "yourapikey"

cc_eth_staking(start = "2022-01-01", end = "2023-01-01", api_key = yourapikey)
cc_eth_staking(start = "2021-01-01", end = "2022-01-01", api_key = yourapikey)
cc_eth_staking(start = "2023-01-01", end = "2023-01-19", api_key = yourapikey)

}
