% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_exchange_historical.R
\name{cc_exchange_historical}
\alias{cc_exchange_historical}
\title{cc_exchange_historical}
\usage{
cc_exchange_historical(
  symbol = NULL,
  start = NULL,
  end = NULL,
  exchange = NULL,
  currency = "USD",
  interval = c("daily", "hourly"),
  api_key = NULL
)
}
\arguments{
\item{symbol}{character, the symbol of interest.}

\item{start}{character or Date, the start date for importation.}

\item{end}{character or Date, the end date for importation.}

\item{exchange}{character, the exchange in which retrieve the information, if NULL it will be returned a mean between all the exchanges and denoted with "global".}

\item{currency}{character, the currency in which convert the output.}

\item{interval}{character, the frequency of the data, it can be "daily", "hourly".}

\item{api_key}{character}
}
\value{
a tibble, with 7 columns (Date, Symbol, Currency, Exchange, VolumeTo, VolumeFrom, Volume)
}
\description{
cc_exchange_historical
}
