% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_options.R
\name{cc_exchange_options}
\alias{cc_exchange_options}
\title{cc_exchage_options}
\usage{
cc_exchange_options()
}
\value{
a character vector with all exchanges
}
\description{
visualize all exchanges availables.
}
