% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_general_info.R
\name{cc_general_info}
\alias{cc_general_info}
\title{Cryptocompare General Info}
\usage{
cc_general_info(
  info = "coins",
  currency = "USD",
  api_key = NULL,
  verbose = FALSE
)
}
\arguments{
\item{info}{Character. Without an \code{api_key} it is possible to access only to \code{"coins"} and \code{"exchanges"} information.
Using a free api key, it is possible to access to all the endpoints:
\itemize{
\item \code{"blockchain"}: returns a list of \href{https://min-api.cryptocompare.com/documentation?key=Blockchain&cat=blockchainListOfCoins}{all coins} for which we currently get blockchain data from IntoTheBlock.
\item \code{"cards"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allCardsStaticInfoEndpoint}{crypto cards}.
\item \code{"coins"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allCoinsWithContentEndpoint}{coins}.
\item \code{"companies"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allMiningCompaniesStaticInfoEndpoint}{mining companies}.
\item \code{"contracts"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allMiningContractsStaticInfoEndpoint}{mining contracts}.
\item \code{"equipment"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allMiningEquipmentStaticInfoEndpoint}{mining equipment}.
\item \code{"exchanges"}: returns general info and 24h volume for all integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allExchangesStaticInfoEndpoint}{exchanges}.
\item \code{"gambling"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allGamblingStaticInfoEndpoint}{exchanges}.
\item \code{"wallets"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allWalletsStaticInfoEndpoint}{wallets}.
\item \code{"pools"}: returns general info about all the integrated \href{https://min-api.cryptocompare.com/documentation?key=Other&cat=allMiningPoolsStaticInfoEndpoint}{mining pools}.
}}

\item{currency}{Character. Currency in which convert the monetary values.}

\item{api_key}{Character, optional. Api key from cyptocompare.}

\item{verbose}{Logical. If \code{TRUE} it will display warning messages if you do not insert a valid api key for the api-key-only endpoints.}
}
\value{
a tibble
}
\description{
Retrieve the general information, using the cryptocompare Api.
}
\examples{
# Endpoints reachable without an Api Key
cc_general_info("coins")
cc_general_info("exchanges")

# Endpoints reachable only with a valid Api Key
\dontrun{
api_key <- "yourapikey"
cc_general_info("gambling", api_key = api_key)
cc_general_info("wallets", api_key = api_key)
cc_general_info("cards", api_key = api_key)
cc_general_info("contracts", api_key = api_key)
cc_general_info("companies", api_key = api_key)
cc_general_info("equipment", api_key = api_key)
cc_general_info("pools", api_key = api_key)
cc_general_info("blockchain", api_key = api_key)
}
}
