% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_general.R
\name{cc_get_general}
\alias{cc_get_general}
\title{cc_get_general}
\usage{
cc_get_general(search = NULL, currency = "USD", quiet = FALSE,
  api_key = NULL)
}
\arguments{
\item{search}{element to search}

\item{currency}{conversion currency}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}

\item{api_key}{api key to access some endpoints}
}
\value{
tbl
}
\description{
function to obtain generic data on the world of cryptocurrencies, the supported methods concern:

\itemize{
\item \code{"all_coins"}: tbl of information for all cryptocurrency.
\item \code{"all_exchanges"}: tbl of all exchanges.
\item \code{"all_gamblings"}: tbl with all gamblings.
\item \code{"all_wallets"}: tbl with all wallets.
\item \code{"all_contracts"}: tbl with all mining contracts.
\item \code{"all_companies"}: tbl with all mining companies.
\item \code{"all_equipment"}: tbl with all mining equipment.
\item \code{"all_pools"}: tbl with all mining pools.
}
}
\examples{
\donttest{cc_get_general(search = "all_coins")}
\donttest{cc_get_general(search = "all_exchanges")}
\donttest{cc_get_general(search = "all_gamlings", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_wallets", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_cards", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_contracts", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_companies", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_equipment", api_key = yourapikey)}
\donttest{cc_get_general(search = "all_pools", api_key = yourapikey)}

}
\seealso{
\itemize{
\item \code{\link{cc_get_options}}
\item \code{\link{cc_all_coins}}:  tbl of information for all cryptocurrency.
\item \code{\link{cc_all_exchanges}}: tbl of all exchanges.
}
}
