% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_news.R
\name{cc_get_news}
\alias{cc_get_news}
\title{cc_get_news}
\usage{
cc_get_news(search = NULL, category = NULL, from = NULL, to = NULL,
  api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference feed that you want to use to acquire data, see the all_feed function.
If NULL, the news will be imported for all feeds.}

\item{category}{reference categories that you want to use to acquire data, see the all_categories function.
If NULL, the news will be imported for all categories.}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
function to obtain data on the news available with the cryptocomp API.
It is possible to search all the news items in a time interval,
or filter them by entering the parameter feeds (the source),
or the parameter category .
It is recommended to check the categories and feeds with the
\code{cc_category_option},
\code{cc_feed_option}, functions.
}
