% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_options.R
\name{cc_get_options}
\alias{cc_get_options}
\title{cc_get_options}
\usage{
cc_get_options()
}
\value{
a character vector with all option for the function \code{\link{cc_get}}
}
\description{
visualize all get options availables.
}
