% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_social.R
\name{cc_get_social}
\alias{cc_get_social}
\title{cc_get_social}
\usage{
cc_get_social(search = NULL, interval = "day", from = NULL,
  to = NULL, api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference symbol that you want to use to acquire data, see the all_symbol function.}

\item{interval}{temporal interval between the data can be:
\itemize{
\item \code{"day"}: daily data  (with "from", default if missing 2008-01-01)
\item \code{"hour"}: horuly data  (with "from", default if missing 2014-01-01)
}}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
a tibble
}
\description{
function for importing social data from \href{https://www.cryptocompare.com}{Cryptocompare}.
}
\examples{
\donttest{cc_get_social(c("BTC","XMR"), interval = "day", api_key = "yourapikey")}
\donttest{cc_get_social(c("BTC","XMR"), interval = "hour", api_key = "yourapikey")}
}
