% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_stock.R
\name{cc_get_stock}
\alias{cc_get_stock}
\title{cc_get_stock}
\usage{
cc_get_stock(search = NULL, interval = "day", exchange = NULL,
  currency = "USD", from = NULL, to = NULL, api_key = NULL,
  quiet = FALSE)
}
\arguments{
\item{search}{reference symbol that you want to use to acquire data, see the all_symbol function.}

\item{interval}{temporal interval between the data can be:
\itemize{
\item \code{"day"}: daily data  (with "from", default if missing 2008-01-01)
\item \code{"hour"}: horuly data  (with "from", default if missing 2014-01-01)
\item \code{"minute"}: inute data for the last 7 days.\preformatted{        }
}}

\item{exchange}{reference exchange that you want to use to acquire data, see the all_exchanges function.
If NULL, the data will be imported for all exchanges.}

\item{currency}{reference currency that you want to use. DEFAUL = "USD"}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
Function to obtain stock data of all cryptocurrency present cryptocompare's Api.
It is possible to obtain data for more than one cryptocurrency per call,
and it is also possible to filter the data using a reference exchange \code{\link{cc_exchange_options}}.
}
