% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_orderbook.R
\name{cc_L1_order_book}
\alias{cc_L1_order_book}
\title{Order Book L1 Top}
\usage{
cc_L1_order_book(symbol, 
                 currency, 
                 exchange = NULL, 
                 api_key = NULL, 
                 quiet = FALSE)
}
\arguments{
\item{symbol}{Character,}

\item{currency}{Character}

\item{exchange}{Character}

\item{api_key}{Character, mandatory.}

\item{quiet}{Logical}
}
\description{
Returns latest order book Level 1 bid/ask values for the specified symbol, currency and exchange.
}
\details{
It is possible to access to this endpoint with an \code{api_key}.
}
