% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_market_info.R
\name{cc_maket_info}
\alias{cc_maket_info}
\title{cryptocompare_api}
\usage{
cc_maket_info(
  info = "coins",
  currency = "USD",
  api_key = NULL,
  verbose = FALSE
)
}
\arguments{
\item{info}{character, the info to be obtained. Without a valid api_key you can access only to the "coins" and "exchanges" endpoints.
Using a free api key, obtained from the website, you can access also to the endpoints:
\itemize{
\item "blockchain", "cards", "companies", "contracts", "equipment", "gambling", "wallets", and "pools".
}}

\item{currency}{character, the currency in which convert the monetary values, if present.}

\item{api_key}{character, a valid api key from cyptocompare.com.}

\item{verbose}{logical, if TRUE it will display warning messages if you do not insert a valid api key for the api-key-only endpoints.}
}
\value{
a tibble, the dimensions of the output depend on the "info" parameters.
}
\description{
function to retrieve the "general informations", using the cryptocompare Api.
}
\details{

}
\examples{
# Endpoints reachable without an Api Key
cc_maket_info("coins")
cc_maket_info("exchanges")

# Endpoints reachable only with a valid Api Key
api_key <- "yourapikey"

cc_maket_info("gambling", api_key = api_key)
cc_maket_info("wallets", api_key = api_key)
cc_maket_info("cards", api_key = api_key)
cc_maket_info("contracts", api_key = api_key)
cc_maket_info("companies", api_key = api_key)
cc_maket_info("equipment", api_key = api_key)
cc_maket_info("pools", api_key = api_key)
cc_maket_info("blockchain", api_key = api_key)

}
