% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_news_categories}
\alias{cc_news_categories}
\title{cc_news_categories}
\usage{
cc_news_categories(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
a tibble
}
\description{
retrieve information about all news categories
}
\examples{
\donttest{cc_news_categories()}
}
\seealso{
\code{\link{cc_get_news}}
\code{\link{cc_get}}
}
