% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_news_feeds}
\alias{cc_news_feeds}
\title{cc_news_feeds}
\usage{
cc_news_feeds(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. When TRUE function evalueates without displaying customary messages}
}
\value{
a tibble
}
\description{
retrieve information about all news feeds
}
\examples{
\donttest{cc_news_feeds()}
}
\seealso{
\code{\link{cc_get_news}}
\code{\link{cc_get}}
}
