% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_orderbook.R
\name{cc_order_book_L2}
\alias{cc_order_book_L2}
\title{Order Book L2 Snapshot}
\usage{
cc_order_book_L2(api_key = NULL)
}
\arguments{
\item{api_key}{Character, mandatory.}
}
\description{
Returns latest order book Level 2 data snapshot for the requested exchange.
}
\details{
It is possible to access to this endpoint with an \code{api_key}.
}
