% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_orderbook.R
\name{cc_orderbook_exchange}
\alias{cc_orderbook_exchange}
\title{Exchanges With Order Book Data}
\usage{
cc_orderbook_exchange(api_key = NULL)
}
\arguments{
\item{api_key}{Character, mandatory.}
}
\description{
Returns all the exchanges we have order book level 2 data.
}
\details{
It is possible to access to this endpoint with an \code{api_key}.
}
