% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_rate_limit}
\alias{cc_rate_limit}
\title{cc_rate_limit}
\usage{
cc_rate_limit(api_key = NULL)
}
\arguments{
\item{api_key}{cryptocompare api key.}
}
\value{
a tibble with information about rate limit
}
\description{
check the rate limit for the month.
}
\examples{
# generic user
\donttest{cc_rate_limit()}
# authenticate user
\donttest{cc_rate_limit(api_key = yourapikey )}

}
