% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_search}
\alias{cc_search}
\title{cc_search}
\usage{
cc_search(x, search = "id")
}
\arguments{
\item{x}{character or NULL, if it is a character it will be setted as api key just if api_key slot in cc_env is empty,
otherwise it will return the value of the api key if the slot is not empty, or NULL if x is NULL and the slot is empty.}

\item{search}{the type of research you want to use, can be :
\itemize{
\item \code{"id"}: for converting a symbol to  id.
\item \code{"symbol"}: for searching a symbol.
\item \code{"exchange"}: for searching an echange.
\item \code{"feed"}:  for searching a news feed.
\item \code{"category"}: for searching a news category.
}}
}
\value{
api_key or NULL
}
\description{
you can use this command to set the api_key global option.
}
\examples{
\donttest{cc_search("BTC", search = "id")}
\donttest{cc_search("BTC", search = "symbol")}
\donttest{cc_search("Coinbase", search = "exchange")}
\donttest{cc_search("yahoo", search = "feed")}

}
\seealso{
\code{\link{cc_set_env}}
}
