% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare.R
\name{cc_set_api_key}
\alias{cc_set_api_key}
\alias{cc_api_key}
\title{cc_set_api_key}
\usage{
cc_set_api_key(api_key = NULL, quiet = FALSE)

cc_api_key(quiet = FALSE)
}
\arguments{
\item{api_key}{character}

\item{verbose}{logical, if TRUE the function display a warning if the Api Key was not found!}
}
\value{
a character, if the api key was added in the "cryptocompare" environment, otherwise NULL.
}
\description{
add a personal api key to the "cryptocompare" environment in such a way to avoid to inserting it manually in every function's call where it is needed.

function to use the api key saved in the "cryptocompare" environment. In order to add it you can use the function \code{cc_set_api_key}
}
\examples{
# Add the api key to the "cryptocompare" environment
yourapikey <- "yourapikey"
cc_set_api_key(api_key = yourapikey)

}
