% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_set_env}
\alias{cc_set_env}
\title{cc_set_env}
\usage{
cc_set_env(api_key = NULL, quiet = FALSE)
}
\arguments{
\item{api_key}{an api key will be added into the environment in order to not type every time the api_key as argument}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\description{
: setting the elements for automatic resarch of inputs in the \code{\link{cc_get}} function.
Allows to do not need to call the api every time but just the first one.
}
\examples{
\donttest{cc_set_env() }
\donttest{cc_set_env(api_key = yourapikey) }
}
\seealso{
\code{\link{cc_clean_env}}
\code{\link{cc_use_key}}
}
