% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyptocompare.R
\name{cc_set_environment}
\alias{cc_set_environment}
\title{cc_set_environment}
\usage{
cc_set_environment()
}
\description{
initialize the environment called "cryptocompare" in which are saved information such as the coins ID, the api key and the exchanges names.
}
