% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_social_historical.R
\name{cc_social_historical}
\alias{cc_social_historical}
\title{cc_social_historical}
\usage{
cc_social_historical(
  symbol = NULL,
  start = NULL,
  end = NULL,
  interval = c("daily", "hourly"),
  api_key = NULL
)
}
\arguments{
\item{symbol}{character, the symbol of interest.}

\item{start}{character or Date, the start date for importation.}

\item{end}{character or Date, the end date for importation.}

\item{interval}{character, the frequency of the data, it can be "daily" or "hourly".}

\item{api_key}{character}
}
\value{
a tibble
}
\description{
cc_social_historical
}
