% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyptocompare.R
\name{cc_symbol_id}
\alias{cc_symbol_id}
\title{cc_symbol_id}
\usage{
cc_symbol_id(symbol = NULL)
}
\arguments{
\item{symbol}{character, the symbol for which we would like to know the Id.}
}
\value{
a character, the symbol ID, if it was found in the "cryptocompare$coins_Id" dataset, otherwise NULL.
}
\description{
retrieve the Id using the symbol, useful for uniforming the social function with the others.
}
\examples{
cc_symbol_id(symbol = "BTC")
cc_symbol_id(symbol = "ETH")

}
