% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_symbol_price.R
\name{cc_symbol_price}
\alias{cc_symbol_price}
\title{cc_symbol_price}
\usage{
cc_symbol_price(
  symbol = NULL,
  exchange = NULL,
  currency = "USD",
  api_key = NULL
)
}
\arguments{
\item{symbol}{A string or character vector containing the symbols of interest. (e.g. "BTC", "ETH", "BNB",...)}

\item{exchange}{A string or character vector containing the exchange to use, if NULL it will be returned a mean between all the exchanges and denoted with "global".}

\item{currency}{A string or character vector containing the currency in which convert the output (e.g. "USD", "USDT", "BNB", "EUR",...).}

\item{api_key}{character}
}
\value{
Return a \code{tibble} object.
}
\description{
function for obtaining the last price for a symbol or a vector of symbols.
}
\examples{
# Mean Price for the Pair BTC-USD between the exchanges
cc_symbol_price(symbol = "BTC", exchange = NULL, currency = "USD")

# Mean Price for the Pair BTC-USD and ETH-USD between the exchanges
cc_symbol_price(symbol = c("BTC", "ETH"), exchange = NULL, currency = c"USD")

# Price for the Pair BTC-USDT, ETH-USDT, BTC-BUSD, ETH-BUSD on Binance
cc_symbol_price(symbol = c("BTC", "ETH"),
                exchange = "Binance",
                currency = c("USDT", "BUSD"))
}
