% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set_env.R
\name{cc_use_key}
\alias{cc_use_key}
\title{cc_use_key}
\usage{
cc_use_key(x = NULL)
}
\arguments{
\item{x}{character or NULL, if it is a character it will be setted as api key just if api_key slot in cc_env is empty,
otherwise it will return the value of the api key if the slot is not empty, or NULL if x is NULL and the slot is empty.}
}
\value{
api_key or NULL
}
\description{
you can use this command to set the api_key global option.
}
\examples{
\donttest{cc_use_key(x = yourapikey)}
}
\seealso{
\code{\link{cc_set_env}}
}
