% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare.R
\name{cryptocompare}
\alias{cryptocompare}
\title{cryptocompare}
\usage{
cryptocompare(search = NULL, method = "stock", interval = "day",
  exchange = NULL, currency = "USD", from = NULL, to = NULL,
  categories = NULL, lang = "EN", api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference element to search.}

\item{method}{the method that you want to use available: "stock","exchange","social","general","news","rate_limit",}

\item{interval}{temporal interval between the data  ("day", "hour", "minute"). DEFAULT "day"}

\item{exchange}{reference exchange that you want to use to acquire data, see the \code{all_exchanges} function.
If NULL, the data will be imported for all exchanges.}

\item{currency}{reference currency that you want to use. DEFAUL = "USD"}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{categories}{reference categories for news method,  see the \code{all_categories} function.
If NULL, the news will be imported for all categories.}

\item{lang}{reference languages for news method, DEFAULT = "EN".}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\description{
cryptocompare generic function, see the vignette C01 - Importing Data with cryptocompareR for more details.
}
\examples{


\dontrun{cryptocompare( search = c("BTC","ETH", "XRP"),method = "stock",
         from = "2019-01-01", to = "2019-06-01", quiet = TRUE)}

\dontrun{cryptocompare(search = c("BTC","ETH", "XRP"), method = "exchange",
          exchange = c("Coinbase","Yobit"), quiet = TRUE)}

\dontrun{ cryptocompare(search = c("BTC", "ETH"), method = "social", api_key = yourapikey,
           from = "2019-06-01", to = "2019-11-10", quiet = T)}


\dontrun{cryptocompare(method = "news", from = "2019-11-01", to = "2019-11-10", quiet = T)}

}
