% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyptocompare_exchange.R
\name{cryptocompare_exchange}
\alias{cryptocompare_exchange}
\title{cryptocompare_exchange}
\usage{
cryptocompare_exchange(search = NULL, interval = "day",
  exchange = NULL, currency = "USD", from = NULL, to = NULL,
  api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference symbol that you want to use to acquire data, see the all_symbol function.
If NULL, the imported data refer to market data (all_symbols).}

\item{interval}{temporal interval between the data  ("day", "hour" ). DEFAULT "day"}

\item{exchange}{reference exchange that you want to use to acquire data, see the all_exchanges function.
If NULL,the imported data refer to market data (all exchanges).}

\item{currency}{reference currency that you want to use. DEFAUL = "USD"}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
function to get the data from exchange from cryptocompare api.
It is a fairly versatile function: if only the search parameter is specified,
it will return the exchange data relating to a specific cryptocurrency for all exchanges;
if only the exchange is specified the exchange,
data refer to all the cryptocurrencies exchanged in that exchange;
finally if the search param == "all", and the exchange is not specified, the data will refer to
the entire market (all cryptocurrencies and all exchanges).
Obviously, by specifying both the symbol and the exchange,
the data will be specific to the two parameters entered.
}
\examples{
\dontrun{cryptocompare_exchange(search = c("BTC","ETH"), from = "2019-06-01")}
\dontrun{cryptocompare_exchange(search = "BTC" , exchange = c("Bitstamp","BitTrex"))}
\dontrun{cryptocompare_exchange(search = c("BTC","ETH"), exchange = "Coinbase" )}

}
