% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_news.R
\name{cryptocompare_news}
\alias{cryptocompare_news}
\title{cryptocompare_news}
\usage{
cryptocompare_news(search = NULL, categories = NULL, from = NULL,
  to = NULL, lang = "EN", api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference feed that you want to use to acquire data, see the all_feed function.
If NULL, the news will be imported for all feeds.}

\item{categories}{reference categories that you want to use to acquire data, see the all_categories function.
If NULL, the news will be imported for all categories.}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{lang}{reference languages, DEFAULT = "EN".}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
function to obtain data on the news available with the cryptocomp API.
It is possible to search all the news items in a time interval,
or filter them by entering the parameter feeds (the source),
or the parameter categories (the category).
It is recommended to check the categories and feeds with the
\code{all_categories},
\code{all_feeds}, functions.
}
