% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_news.R
\name{cryptocompare_news_categories}
\alias{cryptocompare_news_categories}
\title{cryptocompare_news_categories}
\usage{
cryptocompare_news_categories(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
retrieve information about all news categories
}
\examples{
\dontrun{cryptocompare_news_categories()}
}
