% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_news.R
\name{cryptocompare_news_feeds}
\alias{cryptocompare_news_feeds}
\title{cryptocompare_news_feeds}
\usage{
cryptocompare_news_feeds(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. When TRUE function evalueates without displaying customary messages}
}
\value{
tbl
}
\description{
retrieve information about all news feeds
}
\examples{
\dontrun{cryptocompare_news_feeds()}
}
