% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_general.R
\name{cryptocompare_rate_limit}
\alias{cryptocompare_rate_limit}
\title{cryptocompare_rate_limit}
\usage{
cryptocompare_rate_limit()
}
\description{
check the rate limit for the month.
}
\examples{
\dontrun{cryptocompare_rate_limit()}
}
