% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_social.R
\name{cryptocompare_social}
\alias{cryptocompare_social}
\title{cryptocompare_social}
\usage{
cryptocompare_social(search = NULL, interval = "day", from = NULL,
  to = NULL, api_key = NULL, quiet = FALSE)
}
\arguments{
\item{search}{reference symbol that you want to use to acquire data, see the all_symbol function.}

\item{interval}{temporal interval between the data  ("day", "hour"). DEFAULT "day"}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
""
}
