% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_stock.R
\name{cryptocompare_stock}
\alias{cryptocompare_stock}
\title{cryptocompare_stock}
\usage{
cryptocompare_stock(search = NULL, interval = "day", exchange = NULL,
  currency = "USD", from = NULL, to = NULL, api_key = NULL,
  quiet = FALSE)
}
\arguments{
\item{search}{reference symbol that you want to use to acquire data, see the all_symbol function.}

\item{interval}{temporal interval between the data  ("day", "hour", "minute"). DEFAULT "day"}

\item{exchange}{reference exchange that you want to use to acquire data, see the all_exchanges function.
If NULL, the data will be imported for all exchanges.}

\item{currency}{reference currency that you want to use. DEFAUL = "USD"}

\item{from}{start date for import.}

\item{to}{end date for import.}

\item{api_key}{cryptocompare api key.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tbl
}
\description{
Function to obtain stock data of all cryptocurrency present cryptocompare's Api.
It is possible to obtain data for more than one cryptocurrency per call,
and it is also possible to filter the data using a reference exchange.
The time interval can be:
daily with data from 2008-01-01,
hourly with data from 2014-01-01
and minute data for the last 7 days.
}
\examples{
\dontrun{cryptocompare_stock(search = c("XMR","BTC","XRP"))}
\dontrun{cryptocompare_stock("XRP", interval = "hour", exchange = "Coinbase")}
\dontrun{cryptocompare_stock(c("BTC","ETH"), exchange = c("Coinbase","LocalBitcoins"))}

}
