% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_averages.R
\name{EMA}
\alias{EMA}
\title{SMA}
\usage{
EMA(x = NULL, n = 10, smooth = 2, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{n}{integer, the number of length of the moving average.}

\item{smooth}{numeric, a parameter to use for smoothing the moving average.}

\item{na.rm}{logical,}
}
\value{
a vector of the same length of \code{x}.
}
\description{
compute the Exponential Moving Average of a vector.
}
