% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_tidy_list.R
\name{equalize}
\alias{equalize}
\title{equalize}
\usage{
equalize(x = NULL, y = NULL, names = NULL)
}
\arguments{
\item{x}{: a vector}

\item{y}{: a vector}

\item{names}{: if you want to set new names c( "colA", "colB" )}
}
\value{
a tibble with x and y
}
\description{
: equalize two vector of different lengths, it repeat the y-element n-times, with n
equal to the lenght of the x-vector
}
\examples{
\dontrun{equalize(x = c("a", "b"), y = c("d", "e"))}

}
