% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_date.R
\name{format_date}
\alias{format_date}
\title{format_date}
\usage{
format_date(x)
}
\arguments{
\item{x}{a date}
}
\value{
data in character format %Y-%m-%d
}
\description{
a function for formatting a date format %d-%m-%Y
}
\examples{

 format_date("22-11-2019")
 format_date("2019-11-22 19:20:05")

}
