% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_tidy_list.R
\name{map_na}
\alias{map_na}
\title{map_na}
\usage{
map_na(.l, nested = FALSE)
}
\arguments{
\item{.l}{: a list of elements}

\item{nested}{Default FALSE, indicate if you want to map the function in a nested list(list of lists) or in a simple list.}
}
\value{
a list with
}
\description{
: function that iterate over a list the condition na_if ,
specified in the pat argument the function will also perform a substitution of empty elements,
and NULL elements with NA
}
