% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_tidy_list.R
\name{map_tidy}
\alias{map_tidy}
\title{map_tidy}
\usage{
map_tidy(.l = NULL, binded = FALSE)
}
\arguments{
\item{.l}{: a list of list}

\item{binded}{logical, FALSE, indicate if apply bind_rows to reduce the list in a tibble or not.}
}
\value{
a tibble
}
\description{
: mapped version of tidy_list
}
\examples{
 \dontrun{map_tidy(.l = a list of lists)}
}
