% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_env.R
\name{new_cryptocompare_env}
\alias{new_cryptocompare_env}
\title{new_cryptocompare_env}
\usage{
new_cryptocompare_env(quiet = FALSE)
}
\arguments{
\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\description{
: create a new search envirnoment if it doesn't exist
}
