% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_averages.R
\name{SMA}
\alias{SMA}
\title{SMA}
\usage{
SMA(x = NULL, n = 10, .fun = mean, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{n}{integer, the number of length of the moving average.}

\item{.fun}{function, the function to use, for example \code{mean}, \code{var}, \code{sd}, ...}

\item{na.rm}{logical,}
}
\value{
a vector of the same length of \code{x}.
}
\description{
compute the Simple Moving Average of a vector.
}
