% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_tidy_list.R
\name{tidy_list}
\alias{tidy_list}
\title{tidy_list}
\usage{
tidy_list(.l = NULL, quiet = FALSE)
}
\arguments{
\item{.l}{: a list of elements}

\item{quiet}{: indicate if display messages or not}
}
\value{
a tibble
}
\description{
: function for binding a list to dataframe safely the function will also perform a substitution
of empty elements, and NULL elements with NA.
}
\examples{
\dontrun{tidy_list(.l = a list)}
}
