% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_date.R
\name{unix_date}
\alias{unix_date}
\title{unix_date}
\usage{
unix_date(x, from = TRUE, milliseconds = FALSE)
}
\arguments{
\item{x}{a character vector}

\item{from}{if:
- true (default): conversion from unix to date
- false         : conversion from date to unix}

\item{milliseconds}{if:
- true          : conversion in milliseconds}
}
\value{
date converted from or to unix format / milliseconds
}
\description{
a function that convert a date in unix format, from unix to date, or from date/unix to unix in milliseconds.
}
\examples{
 unix_date("2019-11-22 19:20:05", from = FALSE)
 unix_date( "2019-04-01" , from = FALSE, milliseconds = FALSE)
 unix_date( "2019-04-01" , from = FALSE, milliseconds = TRUE)
 unix_date("1574446805",from = TRUE, milliseconds = TRUE)
 unix_date("1574446805",from = TRUE)

}
