% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_unix.R
\name{as_unix}
\alias{as_unix}
\title{as_unix}
\usage{
as_unix(x = NULL, milliseconds = FALSE, origin = "1970-01-01")
}
\arguments{
\item{x}{character or Date, in the format "yyyy-mm-dd" or "dd-mm-yyyy". In alternative it is possible to specify the Date also with the time in the format "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss"}

\item{milliseconds}{logical, if TRUE the output will be converted in milliseconds, otherwise it will be in seconds.}

\item{origin}{character denoting the origin date.}
}
\value{
integer character
}
\description{
This function convert a Date into Unix format: an integer denoting the number of seconds from the origin date, usually "1970-01-01".
}
\examples{
as_unix("2023-01-01")
as_unix("01-01-2023")
as_unix("2023-01-01 10:23:23")
as_unix("01-01-2023 10:23:23")
}
