% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_options.R
\name{cc_all_exchanges}
\alias{cc_all_exchanges}
\title{cc_all_exchanges}
\usage{
cc_all_exchanges()
}
\value{
a tbl with info for all exchanges available
}
\description{
visualize all exchanges information availables.
}
