% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare.R
\name{cc_all_exchanges}
\alias{cc_all_exchanges}
\title{Available Exchanges}
\usage{
cc_all_exchanges()
}
\value{
\code{tibble}
}
\description{
Retrieve the names of all the exchanges available in cryptocompare Api.
}
\examples{
cc_all_exchanges()

}
