% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_eth_staking.R
\name{cc_eth_staking}
\alias{cc_eth_staking}
\title{Staking Rate Historical Day}
\usage{
cc_eth_staking(start = NULL, end = NULL, api_key = NULL)
}
\arguments{
\item{start}{character or Date, the starting date for importation.}

\item{end}{character or Date, the end date for importation.}

\item{api_key}{character}
}
\value{
tibble
}
\description{
Ethereum staking interest rate, from when the Beacon chain was active.
}
\details{
The Ether staking rate is powered by Attestant, a company dedicated to the business of staking.
They provide a non-custodial Ethereum 2 managed staking service, giving customers the ability to stake
their Ether using Attestant’s infrastructure while always retaining full control of their assets.
You can only use this endpoint with a valid api_key.
The Ether rate is generally updated daily at 11:00 am UTC for the previous day.
}
\examples{
# Api key is required
\dontrun{
yourapikey <- "yourapikey"
cc_eth_staking(start = "2022-01-01", end = "2023-01-01", api_key = yourapikey)
cc_eth_staking(start = "2021-01-01", end = "2022-01-01", api_key = yourapikey)
cc_eth_staking(start = "2023-01-01", end = "2023-01-19", api_key = yourapikey)
}

}
