% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_rate_limit.R
\name{cc_rate_limit}
\alias{cc_rate_limit}
\title{Cryptocompare Rate limit}
\usage{
cc_rate_limit(hour = FALSE)
}
\arguments{
\item{hour}{Logical. Default is \code{FALSE}. If \code{TRUE} will return the number of calls done and left in the current hour.}
}
\description{
Find out how many calls you have left in the current month, day, hour, minute and second
}
\examples{

cc_rate_limit(hour = FALSE)
cc_rate_limit(hour = TRUE)

}
