% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare.R
\name{cc_symbol_id}
\alias{cc_symbol_id}
\title{Cryptocompare Symbol Id}
\usage{
cc_symbol_id(symbol = NULL)
}
\arguments{
\item{symbol}{Character.}
}
\value{
The symbol id if was found, otherwise \code{NULL}.
}
\description{
Retrieve the id for a symbol.
}
\examples{
cc_symbol_id(symbol = "BTC")
cc_symbol_id(symbol = "ETH")

}
