% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptocompare_api.R
\name{cryptocompare_api}
\alias{cryptocompare_api}
\title{GET call to cryptocompare API}
\usage{
cryptocompare_api(path = NULL, query = NULL, config = list())
}
\arguments{
\item{path}{Character vector. API path, \code{NULL} or \code{NA} elements will be excluded.}

\item{query}{Named list. Query parameters for the API call, \code{NULL} or \code{NA} elements will be excluded.}

\item{config}{Configuration parameter \code{config}. See the function \code{\link[httr]{GET}}.}
}
\description{
Execute a GET call to cryptocompare API.
}
\examples{
api_path = c("data", "exchange", "histoday")
api_query = list(tsym = "USD", e = "Binance",
                 toTs = as_unix("2023-01-01"),
                 limit = 2000, api_key = NULL)
cryptocompare_api(path = api_path, query = api_query)

}
